pyproject.toml
[project]
name = "fastapi-pydantic-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1,<1.0",
    "httpx>=0.28.1,<1.0",
    "llama-index-core>=0.14.18,<1.0",
    "llama-index-workflows>=2.17.0,<3.0",
    "opentelemetry-instrumentation-fastapi>=0.51b0,<1.0",
    "pydantic-ai-slim[logfire,openai]>=1.70.0,<2.0",
    "pydantic-settings>=2.13.1,<3.0",
    "slowapi>=0.1.9,<1.0",
    "uvicorn[standard]>=0.42.0,<1.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2,<10.0",
    "pytest-asyncio>=1.3.0,<2.0",
    "pytest-cov>=7.0.0,<8.0",
    "ruff>=0.15.7,<1.0",
    "ty>=0.0.24,<0.1.0",
]

# ============================================================================
# Ruff Configuration - Linter & Formatter
# ============================================================================
[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "D",    # pydocstyle (docstring)
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "S",    # flake8-bandit (Security)
    "ANN",  # flake8-annotations
    "RUF",  # Ruff-specific rules
    "T20",  # flake8-print
    "PT",   # flake8-pytest-style
    "SIM",  # flake8-simplify
]

ignore = [
    "D104", # Missing docstring in public package
]

[tool.ruff.lint.isort]
force-single-line = true
lines-after-imports = 2

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ["S101", "ANN", "S603", "S607"]

# ============================================================================
# Pytest Configuration
# ============================================================================
[tool.pytest.ini_options]
markers = [
    "docker: marks tests that require Docker to be installed and available (skip in local dev, run in CI)",
]

# ============================================================================
# Ty (Type Checker) Configuration - Strict type checking
# ============================================================================
[tool.ty]

[tool.ty.analysis]
# Respect type: ignore comments in code
respect-type-ignore-comments = true
# Don't replace any imports with Any (empty list = strict mode)
replace-imports-with-any = []
